+++ playerClient/JavaScript/cards.js
@@ -1,4 +1,5 @@
 import * as Main from './main.js';
+
 export class Card{
     constructor(){
