--- playerClient/JavaScript/utilityFunctions.js
@@ -1,14 +1,0 @@
-export function $(el) { return document.getElementById(el) };
-export function getOffset( el ) {
-    var _x = 0;
-    var _y = 0;
-    while( el && !isNaN( el.offsetLeft ) && !isNaN( el.offsetTop ) ) {
-        _x += el.offsetLeft - el.scrollLeft;
-        _y += el.offsetTop - el.scrollTop;
-        el = el.offsetParent;
-    }
-    return { top: _y, left: _x };
-}
-export function getUIHandler(){
-    return UI_Handle;
-}
